[ADDED] logging out, and changed the way token is retrieved into state
Now the user can log out (localStorage, or sessionStorage is cleared) and is redirected to the main gate (app is refreshed and cleared).
Token is now retrieved not in a function inside componentDidMount, as it is async, and shouldn't be in such a lifecycle method. Instead, it is retrieved from the local or session storage inside the constructor before declaring state, and then it is stored in the state.