[FIXED] Single blog view
Now the user can see the single post, upon clicking it, all the data is now provided by the server inside componentdidmount axios call as opposed to getting the data from props as before. The id of the post is sent through URLsearchParams, and then is given to the server so that it can search the database with it.