[ADDED] a limit on the request for posts
there is a limit on the number of posts that the user can see at the begginning. This helps to optimize the application by reducing the payload of the get posts query that doesn't need to send all the posts now, but only the first few.
The posts are sorted so that the newest is not top of the displayed list.