[FIXED] user photo not being updated and links not working
the error was that to fetch the user photo I used the user id from the currently logged in user instead of the user whose profile was viewed. That is now fixed.
Also I fixed the link in the userphoto component (top menu) to be of type 'a' instead of Link because it wasn't working if the user was look at another user's profile and then wanted to jump to his own profile