[FIXED] multiple res in one route error, and security error
instead of seding multiple responses from one route (which generated errors) I send just one response and return where I would have send a response earlier.
The security flaw fixed is that the authorId was taken from the request body which could be used to set likes/dislikes for other users. That's why now the authorId is taken from the auth middleware which takes it from the token.