[ADDED] correct type logic for getting data for likes count
There are 4 options:
like is of type blog and is positive,
like is of type comment and is positive,
like is of type blog and is negative,
like is of type comment and is negative.
All those options are now logically divided and can be controlled via props.
based on this logic, the component sends the adequate axios call