[ADDED] the user profile component now checks if the user is a friend of the logged in user
in componentDidMount (on init) this is information is checked, and based on it the friend button text either offers to remove the friend or to send a request to the friend.
For this to work a new API path has been created. It's called 'checkFriendStatus' and requires the user id and the friend id. Then it searches the database with those credentials and if such connection exists, then it returns true, otherwise it returns false