[FIXED] and modified the friend button so that the text is displayed properly now
I changed the button to a class component, and added an isFriend value to the state, which initially is based on props.isFriend. The change was done to allow the button to controll it's own 'knowledge' of the friend status based on user actions. The parent component state will be affected by the functions passed in props that are triggered when the button is pressed.