[FIXED] a serious api error with sending requests
If someone tried to send a friend request to a specific user, then if that user already had a friend request pending from someone else, the new friend request would not be created and the server would deny sending anything. Now the system is refined, and checks if the user has sent the request previously, and also if the friend-to-be has sent the request to the user. In that way, the api ensures that no duplicate friend requests are created.