[FIXED] edit comment form appearing for every comment when the user pressed the edit button
each comment was now moved to a separate comment component that has options and edit comment form components. In this way pressing edit on one component doesn't make the form appear on another comment, because the state boolean for editing is now in each single comment component, as opposed to being in all comments component.