[ADDED] back to top scroll listener higher order component, and fixed some styling with photos in my activity
the scroll listener has a use effect that starts an event listener for scrolling. If scroll is over 350 then a 'back to top' button appears. This is a higher order component that encompasses the entire 'logged in' part of the App.

In styles I gave the photos in myActivity a constant value to avoid unwarrented behavior