[MODIFIED] conversation model to include conversation type (group/private)
If conversation type is private then the name will always be blank, as it wont be displayed. (In place of the conversation name there will be the name of the other participant). The limit on participants in a private conversation is 2: the user and his friend. No friends may be added to a private conversation, and such conversation cannot be left, instead it is deleted