[FIXED] minor bugs with loading more comments
now the comments array is not reversed but the new comment that was posted is added to the begginning of the array at index = 0.
Also, there was a bug where the limit on the number of comments that could be retrieved was 4 on the server, and locally after loading more comments the limit wasn't increased by 4, but by 2. That mistake has been corrected