[ADDED] front end deleting notification animation and logic
In the newly added removeNotif function there are three parameters:
isFriendRequest, deletingAll, data
Those parameters will be used for axios calls to API regarding deleting different notifications and friendRequests.
In this function also the notification count is decremented to represent the number of still active notifications.
On the front end after pressing the delete icon a new class translatesX the notification until it overflows the parent div (which has overflowX hidden) and with a 600ms delay another class is asigned that changes the display to none, thus making the notification seem 'deleted' for the user