[FIXED] conversations server and front end logic. Menu component joins all socket channels accordingly
Menu component executes a for Each loop on all conversations for given user fetched from the server. The for each loop joins all conversation id websocket rooms on the server and listens for events. If the user is not in the chat component, then on 'message' event the message counter is incremented. If the user is in the chat component then the application reacts only if the user is in the correct conversation id.